requirements.txt
Flask
gspread
oauth2client
pandas
openpyxl
python-dotenv
flask-login
werkzeug
Plotly
